+++ docs/source/HISTORY.rst
@@ -4,6 +4,7 @@ Changelog
 1.3.0b2 (unreleased)
 --------------------
 - added 'create_if_not_existing' parameter to webdav_handle() method
+- added ensuredir() to wrapper base class
 
 1.3.0b1 (2015-10-15)
 --------------------